requirements.txt
packaging==24.1
pillow==10.4.0
pyperclip==1.9.0
pytesseract==0.3.13
